requirements.txt
# Core dependencies
numpy>=1.26.0,<2
opencv-python>=4.8.0
Pillow>=10.0.0

# Deep learning
torch>=2.7.0
torchvision>=0.20.0

# Detection - RF-DETR (Apache 2.0)
rfdetr>=1.2.0

# Segmentation - SAM2/SAM3 (install from GitHub)
# pip install git+https://github.com/facebookresearch/sam2.git
timm>=0.9.0
einops>=0.7.0
huggingface_hub>=0.20.0

# VLM - Qwen support
transformers>=4.40.0
accelerate>=0.25.0
aiohttp>=3.9.0

# API and utilities
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
websockets>=12.0
pydantic>=2.0.0

# Optional: RunPod serverless
# runpod>=1.6.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
